Cleanup IPv6AddressHelper to use Span<T> where possible.
There were two slightly incompatible versions of `IPv6AddressHelper` in
`System.Private.Uri` and `System.Net.Primitives`; causing problems for Mono's
adaption where both need to live in the same `System.dll` assembly.

Moved the entire `System.Net.Primitives` version into `Common` and adjusted
its API to make it usable from `System.Private.Uri`.
