+++ src/System.Net.Primitives/src/System.Net.Primitives.csproj
@@ -39,7 +39,6 @@
     <Compile Include="$(CommonPath)\System\Net\IPv4AddressHelper.Common.cs">
       <Link>System\Net\IPv4AddressHelper.Common.cs</Link>
     </Compile>
-    <Compile Include="System\Net\IPv6AddressHelper.cs" />
     <Compile Include="$(CommonPath)\System\Net\IPv6AddressHelper.Common.cs">
       <Link>System\Net\IPv6AddressHelper.Common.cs</Link>
     </Compile>
