+++ src/System.Net.Primitives/tests/PalTests/System.Net.Primitives.Pal.Tests.csproj
@@ -40,9 +40,6 @@
     <Compile Include="$(CommonPath)\System\Net\IPv4AddressHelper.Common.cs">
       <Link>ProductionCode\System\Net\IPv4AddressHelper.Common.cs</Link>
     </Compile>
-    <Compile Include="..\..\src\System\Net\IPv6AddressHelper.cs">
-      <Link>ProductionCode\System\Net\IPv6AddressHelper.cs</Link>
-    </Compile>
     <Compile Include="$(CommonPath)\System\Net\IPv6AddressHelper.Common.cs">
       <Link>ProductionCode\System\Net\IPv6AddressHelper.Common.cs</Link>
     </Compile>
