+++ src/System.Net.Primitives/tests/UnitTests/Fakes/IPv6AddressHelper.cs
@@ -9,7 +9,7 @@ namespace System
     internal static class IPv6AddressHelper
     {
         internal unsafe static (int longestSequenceStart, int longestSequenceLength) FindCompressionRange(
-            ushort[] numbers, int fromInclusive, int toExclusive) => (-1, -1);
+            ReadOnlySpan<ushort> numbers) => (-1, -1);
         internal unsafe static bool ShouldHaveIpv4Embedded(ushort[] numbers) => false;
         internal unsafe static bool IsValidStrict(char* name, int start, ref int end) => false;
         internal static unsafe bool Parse(ReadOnlySpan<char> ipSpan, ushort* numbers, int start, ref string scopeId) => false;
